+++ packages/eden-react-query/src/implementation/treaty/use-queries.ts
@@ -1,9 +1,14 @@
-import type {
-  EdenClient,
-  EdenRequestParams,
-  InferRouteError,
-  InferRouteOptions,
-  InferRouteOutput,
+import {
+  type EdenClient,
+  type EdenRequestParams,
+  type EmptyToVoid,
+  type ExtractEdenTreatyRouteParams,
+  type ExtractEdenTreatyRouteParamsInput,
+  getPathParam,
+  type InferRouteError,
+  type InferRouteOptions,
+  type InferRouteOutput,
+  parsePathsAndMethod,
 } from '@ap0nia/eden'
 import type {
   QueriesOptions,
@@ -15,12 +20,6 @@ import type {
 import type { AnyElysia, RouteSchema } from 'elysia'
 
 import type { EdenQueryConfig } from '../../config'
-import { parsePathsAndMethod } from '../../integration/internal/parse-paths-and-method'
-import {
-  type ExtractEdenTreatyRouteParams,
-  type ExtractEdenTreatyRouteParamsInput,
-  getPathParam,
-} from '../../integration/internal/path-params'
 import type { EdenQueryBaseOptions } from '../../integration/internal/query-base-options'
 import { type EdenQueryKey, getQueryKey } from '../../integration/internal/query-key'
 import type { UseQueryOptionsForUseQueries } from '../../integration/internal/use-query-options-for-use-queries'
@@ -77,7 +76,7 @@ export type EdenTreatyUseQueriesHook<
   TError = InferRouteError<TRoute>,
   TKey extends QueryKey = EdenQueryKey<TPath>,
 > = (
-  input: {} extends TInput ? void | TInput : TInput,
+  input: EmptyToVoid<TInput>,
   opts?: UseQueryOptionsForUseQueries<TOutput, TInput, TError>,
 ) => UseQueryOptions<TOutput, TError, TOutput, TKey>