+++ packages/eden-react-query/src/integration/hooks/use-infinite-query.ts
@@ -1,4 +1,10 @@
-import type { InferRouteError, InferRouteOptions, InferRouteOutput } from '@ap0nia/eden'
+import type {
+  EmptyToVoid,
+  InferRouteError,
+  InferRouteOptions,
+  InferRouteOutput,
+  ParsedPathAndMethod,
+} from '@ap0nia/eden'
 import {
   type InfiniteData,
   type InfiniteQueryObserverSuccessResult,
@@ -12,7 +18,6 @@ import type { RouteSchema } from 'elysia'
 import { type EdenContextState, useSSRQueryOptionsIfNeeded } from '../../context'
 import type { DistributiveOmit } from '../../utils/types'
 import type { ExtractQueryCursor, ReservedInfiniteQueryKeys } from '../internal/infinite-query'
-import type { ParsedPathAndMethod } from '../internal/parse-paths-and-method'
 import type { EdenQueryBaseOptions } from '../internal/query-base-options'
 import type { WithEdenQueryExtension } from '../internal/query-hook-extension'
 import { getQueryKey } from '../internal/query-key'
@@ -50,7 +55,7 @@ export type EdenUseInfiniteQuery<
   TError = InferRouteError<TRoute>,
   TInfiniteInput = InferRouteOptions<TRoute, ReservedInfiniteQueryKeys>['query'],
 > = (
-  input: {} extends TInfiniteInput ? void | TInfiniteInput : SkipToken,
+  input: EmptyToVoid<TInfiniteInput> | SkipToken,
   options: EdenUseInfiniteQueryOptions<TInput, TOutput, TError>,
 ) => EdenUseInfiniteQueryResult<TOutput, TError, TInput>