+++ packages/eden-svelte-query/src/context.ts
@@ -1,4 +1,9 @@
-import type { EdenClient, EdenClientError, EdenRequestParams } from '@ap0nia/eden'
+import {
+  type EdenClient,
+  type EdenClientError,
+  type EdenRequestParams,
+  parsePathsAndMethod,
+} from '@ap0nia/eden'
 import type {
   CancelOptions,
   FetchQueryOptions,
@@ -20,21 +25,22 @@ import type { AnyElysia } from 'elysia'
 import type { EdenQueryConfig } from './config'
 import type { EdenFetchInfiniteQueryOptions } from './integration/hooks/fetch-infinite'
 import type { EdenFetchQueryOptions } from './integration/hooks/fetch-query'
-import { parsePathsAndMethod } from './integration/internal/parse-paths-and-method'
 import type { EdenMutationKey, EdenQueryKey, EdenQueryType } from './integration/internal/query-key'
 
 export const EDEN_CONTEXT_KEY = Symbol('EDEN_CONTEXT')
 
 export type EdenContextPropsBase<TElysia extends AnyElysia, _TSSRContext = unknown> = {
   /**
-   * The `TRPCClient`
+   * Untyped client for making requests.
    */
   client: EdenClient<TElysia>
 
   /**
-   * @deprecated pass abortOnUnmount to `createTRPCReact` instead
-   * Abort loading query calls when unmounting a component - usually when navigating to a new page
+   * Whether to forward the `signal` from svelte-query to fetch call.
+   *
    * @default false
+   *
+   * @deprecated pass abortOnUnmount to `createTRPCReact` instead
    */
   abortOnUnmount?: boolean
 }