+++ packages/eden/src/links/http-batch-link.ts
@@ -21,12 +21,12 @@ import {
 
 /**
  * @remarks Do not derive this from HTTPLinkOptions, because it breaks the types for some reason...
+ *
+ * @template TTransformer
+ * @todo Maybe check if T['store'][EdenQueryStoreKey] matches a certain interface?
  */
 export type HttpBatchLinkOptions<
   T extends AnyElysia = AnyElysia,
-  /**
-   * @todo Maybe check if T['store'][EdenQueryStoreKey] matches a certain interface?
-   */
   TTransformer = T['store'][typeof EdenQueryStoreKey]['transformer'],
 > = Omit<EdenRequestOptions, 'headers' | 'method' | 'transformer'> & {
   /**
@@ -49,11 +49,15 @@ export type HttpBatchLinkOptions<
     | ((operations: NonEmptyArray<Operation>) => HTTPHeaders | Promise<HTTPHeaders>)
 
   method?: BatchMethod
-} & (TTransformer extends DataTransformerOptions
-    ? { transformer: TTransformer }
-    : {
+} & (TTransformer extends false
+    ? {
         transformer?: DataTransformerOptions
-      })
+      }
+    : TTransformer extends DataTransformerOptions
+      ? { transformer: TTransformer }
+      : {
+          transformer?: DataTransformerOptions
+        })
 
 export type BatchMethod = 'GET' | 'POST'
 
@@ -464,9 +468,9 @@ function createBatchRequester(options: HttpBatchLinkOptions = {}): Requester {
 }
 
 /**
- * @link https://trpc.io/docs/v11/client/links/httpLink
+ * @see https://trpc.io/docs/v11/client/links/httpLink
  */
-export const httpBatchLink = <T extends AnyElysia>(
+export const safeHttpBatchLink = <T extends AnyElysia>(
   options?: HttpBatchLinkOptions<T>,
 ): T['store'][typeof EdenQueryStoreKey]['batch'] extends true | BatchPluginOptions
   ? EdenLink<T>
@@ -476,10 +480,10 @@ export const httpBatchLink = <T extends AnyElysia>(
 }
 
 /**
- * @link https://trpc.io/docs/v11/client/links/httpLink
+ * @see https://trpc.io/docs/v11/client/links/httpLink
  */
-export function unsafeHttpBatchLink<T extends AnyElysia>(
-  options?: HttpBatchLinkOptions<T>,
+export function httpBatchLink<T extends AnyElysia>(
+  options?: HttpBatchLinkOptions<T, false>,
 ): EdenLink<T> {
-  return httpBatchLink(options) as any
+  return safeHttpBatchLink(options as any) as any
 }