+++ packages/eden-next-query/CHANGELOG.md
@@ -1,5 +1,123 @@
 # @ap0nia/eden-next-query
 
+## 1.4.3
+
+### Patch Changes
+
+- Updated dependencies [37ee3c0]
+  - @ap0nia/eden-react-query@1.6.0
+  - @ap0nia/eden@1.6.0
+
+## 1.4.2
+
+### Patch Changes
+
+- Updated dependencies [ef7456e]
+- Updated dependencies [c9dbba8]
+  - @ap0nia/eden-react-query@1.5.0
+  - @ap0nia/eden@1.5.0
+
+## 1.4.1
+
+### Patch Changes
+
+- 17ddc6e: fix(eden): treaty types
+- Updated dependencies [8de6067]
+- Updated dependencies [17ddc6e]
+  - @ap0nia/eden-react-query@1.4.1
+  - @ap0nia/eden@1.4.1
+
+## 1.4.0
+
+### Minor Changes
+
+- c76900c: feat: eden-treaty and updated APIs
+
+### Patch Changes
+
+- Updated dependencies [c76900c]
+  - @ap0nia/eden-react-query@1.4.0
+  - @ap0nia/eden@1.4.0
+
+## 1.3.1
+
+### Patch Changes
+
+- 17b99f2: fix(eden): infinite recursion with http-link
+- Updated dependencies [17b99f2]
+- Updated dependencies [17b99f2]
+  - @ap0nia/eden@1.3.1
+  - @ap0nia/eden-react-query@1.3.1
+
+## 1.3.0
+
+### Minor Changes
+
+- 3e231fd: refactor(eden): expose server plugins in different entrypoint file
+
+### Patch Changes
+
+- Updated dependencies [3e231fd]
+  - @ap0nia/eden-react-query@1.3.0
+  - @ap0nia/eden@1.3.0
+
+## 1.2.1
+
+### Patch Changes
+
+- 25151e7: refactor(eden): make eden plugin unsafe by default
+- Updated dependencies [25151e7]
+  - @ap0nia/eden@1.2.1
+  - @ap0nia/eden-react-query@1.2.1
+
+## 1.2.0
+
+### Minor Changes
+
+- 456df95: fix(eden): updated error checker
+
+### Patch Changes
+
+- Updated dependencies [456df95]
+  - @ap0nia/eden@1.2.0
+  - @ap0nia/eden-react-query@1.2.0
+
+## 1.1.0
+
+### Minor Changes
+
+- 5ff7808: refactor: default to unsafe http and http-batch links
+
+### Patch Changes
+
+- Updated dependencies [5ff7808]
+  - @ap0nia/eden-react-query@1.1.0
+  - @ap0nia/eden@1.1.0
+
+## 1.0.15
+
+### Patch Changes
+
+- Updated dependencies [47d786f]
+  - @ap0nia/eden-react-query@1.0.15
+
+## 1.0.14
+
+### Patch Changes
+
+- Updated dependencies [17424ac]
+  - @ap0nia/eden@1.0.13
+  - @ap0nia/eden-react-query@1.0.14
+
+## 1.0.13
+
+### Patch Changes
+
+- ec1fa09: fix(eden): do not initialize plugins multiple times
+- Updated dependencies [ec1fa09]
+  - @ap0nia/eden@1.0.12
+  - @ap0nia/eden-react-query@1.0.13
+
 ## 1.0.12
 
 ### Patch Changes